Add real type annotations for result4
This one was a bit trickier since the C module function can actually
return a 4-tuple or a 6-tuple and the comments in Modules/message.c
were misleading.

Simplify the C module by always returning a 6-tuple (unless a polling
operation is requested, in which case the return value will be a simple
None).

That makes it more similar to the Python wrapper counterpart.