[python-ldap] Type fixes for Lib/ldap/cidict.py
First, __contains__() can be called with anything as the key, but only
supports str keys (due to the key.lower() call), so make sure that the
key is actually a str.

Second, strlist_union() is described as returning a list, but actually
returns an iterable, so make it return a list.

@@ -134,4 +136,4 @@ def strlist_union(a,b):