[python-ldap] Type fixes for Lib/ldap/controls/simple.py
The comments for the ProxyAuthzControl and AuthorizationIdentityResponseControl
classes indicate that they expect/return a str (which makes sense), but then
fail to actually deal with the str properly.

The GetEffectiveRightsControl class has a similar issue (but seems to be a stub
at the moment, the "authzId" parameter should probably be renamed "gerSubject"
and can take much more than just a DN).