[python-ldap] Type fixes for Lib/ldap/syncrepl.py
First, the criticality in class SyncRequestControl can be passed as an
int/bool/etc, but it still makes sense (and helps type checkers) to
make sure that it is actually stored as a bool.

Second, some minor code changes to exclude the possibility that some variables
are not None and to help type checkers understand the type of some objects.

Last, make sure that the syncrepl_get_cookie() function is type-conformant.